delete workout

+++ src/services/app_service.py
@@ -20,6 +20,8 @@ class InvalidDate(Exception):
 class DuplicateWorkoutError(Exception):
     pass
 
+class NoSuchWorkoutError(Exception):
+    pass
 
 class AppService:
 
@@ -102,13 +104,17 @@ class AppService:
             raise InvalidDate(
                 "Päivämäärä väärässä muodossa. Käytä muotoa YYYY-MM-DD") from exception
 
-        self._workout_repository.delete_one_workout(self._user.username, content, date)
+        user_workouts = self.get_user_workouts()
 
-        workouts = self._workout_repository.list_all_user_workouts(
-            self._user.username)
+        found_workout = False
+        for workout in user_workouts:
+            if workout.content == content and workout.date == date:
+                self._workout_repository.delete_one_workout(self._user.username, content, date)
+                found_workout = True
+                break
 
-        return workouts
-
+        if not found_workout:
+            raise NoSuchWorkoutError("Treeniä ei löydy annetulla sisällöllä ja päivämäärällä")
 
 # def modify_workout
